// lib/types.ts
export type Event = {
  id: string
  name: string
  start_datetime: string | null
  end_datetime: string | null
  location_name: string | null
  address: string | null
  description: string | null
  is_all_day: boolean
  type: "in-person" | "virtual" | "hybrid"
  url: string | null
  status: "active" | "draft"
  categories: { category: { id: string; name: string } }[]
  tags: { tag: string }[]
  profile: {
    username: string
    profile_pic_url: string | null
    bio: string | null
  } | null
  school: {
    name: string
    address: string | null
  } | null
  post_images: {
    post_images: { file_path: string }[]
  } | null
}

export type PopularEvent = {
  event_id: string
  event_name: string
  start_datetime: string | null
  tag_count: number
  // We need to fetch the image separately for popular events
  post_images?: { file_path: string }[] | null
}

export const CATEGORIES = ["event", "club", "sport", "deadline", "meeting"] as const